[FIX] Domain 패키지별 설정
