+++ src/main/java/com/crewspace/api/domain/member/Authority.java
@@ -1,4 +1,4 @@
-package com.crewspace.api.domain;
+package com.crewspace.api.domain.member;
 
 public enum Authority {
     ROLE_USER, ROLE_ADMIN
