+++ src/main/java/com/crewspace/api/domain/post/CommunityPost.java
@@ -1,8 +1,7 @@
-package com.crewspace.api.domain;
+package com.crewspace.api.domain.post;
 
+import com.crewspace.api.domain.spaceMember.SpaceMember;
 import javax.persistence.Entity;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @Entity
-public class CommunityPost extends Post{
+public class CommunityPost extends Post {
     @Builder
     public CommunityPost(SpaceMember author, PostCategory postCategory, String description) {
         super(author, postCategory, description);
